+++ src/forgegate/live_status/__init__.py
@@ -1,8 +1,9 @@
-from forgegate.live_status.models import LiveSourceStatus, LiveStatusPage
+from forgegate.live_status.models import LiveReportedIssue, LiveSourceStatus, LiveStatusPage
 from forgegate.live_status.provider import DisabledLiveStatusProvider, LiveStatusProvider
 
 __all__ = [
     "DisabledLiveStatusProvider",
+    "LiveReportedIssue",
     "LiveSourceStatus",
     "LiveStatusPage",
     "LiveStatusProvider",